requirements.txt
certifi==2018.4.16
future==0.16.0
numpy==1.14.3
psycopg2==2.7.4
python-telegram-bot==10.0.2
SQLAlchemy==1.2.7
